entry point: add option to disable it
This adds a NO_RUN_APPLICATION macro which, when defined, disables the
RunApplication entry point so the user can use WinMain, main, etc directly.

It should be noted that this patch may introduce false positives during
Windows memory leak checking and if NO_RUN_APPLICATION is defined, the
crash reports facility is disabled.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>
Reported-by: Alastair Lynn <arplynn@gmail.com>