STree: a very simple binary search tree
This is a stupid binary search tree which has no self-balancing features
or nice optimizations. It works well with very random data, but that's
about it. Using this tree in production code is VERY highly discouraged.

The only reason I'm adding it is because it can be a great template for
future binary trees.

(Interesting note: this tree is very much like the binary search tree
that Introversion Software uses in Uplink, Darwinia, and Defcon. The
only difference is that this tree doesn't use recursion so wildly, which
makes this tree slightly safer when it hits a pathological case, such as
insertion/searching/deletion of ordered data.)

Signed-off-by: Steven Noonan <steven@uplinklabs.net>