debug: fix __cxa_demangle on Clang with libc++
In this scenario, cxxabi.h doesn't end up on the include path because
it's part of GCC, not Clang. And there aren't any Clang headers which
declare __cxa_demangle. But libc++abi *does* define the symbol.

So I've just copied the one symbol declaration we use from the header
here, if we're compiling using Clang.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>