DArray/DStack: add options useful for static instantiation
The way DArray is currently designed is optimal for long-term instances
with random inseration/deletion.

The design choices that make this perform well don't help another common
use case: a DArray that is only filled linearly and emptied between uses
(and statically allocated). Maintaining the free index stack is brutally
expensive in that scenario.

This adds a couple options to avoid freeing memory, and even disabling
the stack optimization.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>