trees: implement support for type erasure
Useful for code size reduction when using trees to store disparate
pointer types.

For example instead of:

    BTree<const char *, SomeClass *> tree1;
    BTree<const char *, OtherClass *> tree2;

    SomeClass *ptr1 = tree1.find("foo");
    OtherClass *ptr2 = tree2.find("bar");

you could now use:

    BTree<const char *, void *> tree1;
    BTree<const char *, void *> tree2;

    SomeClass *ptr1 = tree1.find<SomeClass *>("foo");
    OtherClass *ptr2 = tree2.find<OtherClass *>("bar");

The latter avoids a new type instance of the BTree class, and results in
some dramatic code size reductions.

The downside is that type safety is reduced, and it's up to the
developer to ensure they do the right thing.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>